remove dispatch from child components
